@@ -14,21 +14,21 @@ Gem::Specification.new do |s|
     f.match(%r{^(spec|features)/})
   end
 
-  s.test_files    = `git ls-files -- {spec,features}/*`.split("\n")
+  s.test_files = `git ls-files -- {spec,features}/*`.split("\n")
 
-  s.required_ruby_version = '>= 2.1'
+  s.required_ruby_version = '>= 2.2'
 
-  s.add_dependency 'arbre',               '~> 1.0', '>= 1.0.2'
+  s.add_dependency 'arbre', '>= 1.1.1'
   s.add_dependency 'bourbon'
   s.add_dependency 'coffee-rails'
-  s.add_dependency 'formtastic',          '~> 3.1'
+  s.add_dependency 'formtastic', '~> 3.1'
   s.add_dependency 'formtastic_i18n'
   s.add_dependency 'inherited_resources', '~> 1.7'
   s.add_dependency 'jquery-rails'
   s.add_dependency 'jquery-ui-rails'
-  s.add_dependency 'kaminari',            '>= 0.15', '< 2.0'
-  s.add_dependency 'railties',            '>= 4.0', '< 5.1'
-  s.add_dependency 'ransack',             '~> 1.3'
+  s.add_dependency 'kaminari', '>= 0.15', '< 2.0'
+  s.add_dependency 'railties', '>= 4.2', '< 5.2'
+  s.add_dependency 'ransack', '~> 1.3'
   s.add_dependency 'sass-rails'
-  s.add_dependency 'sprockets',           '< 4.1'
+  s.add_dependency 'sprockets', '< 4.1'
 end
