@@ -4,7 +4,7 @@ end
 
 When /^I add a comment "([^"]*)"$/ do |comment|
   step %{I fill in "active_admin_comment_body" with "#{comment}"}
-  step  %{I press "Add Comment"}
+  step %{I press "Add Comment"}
 end
 
 Given /^a tag with the name "([^"]*)" exists$/ do |tag_name|
