@@ -27,10 +27,6 @@ When /^(?:I )go to (.+)$/ do |page_name|
   visit path_to(page_name)
 end
 
-When /^(?:I )POST to "(.*?)" with params "(.*?)"$/ do |url, params|
-  post "#{url}#{params}"
-end
-
 When /^(?:I )press "([^"]*)"$/ do |button|
   click_button(button)
 end
@@ -72,7 +68,12 @@ end
 
 Then /^the "([^"]*)" checkbox(?: within (.*))? should( not)? be checked$/ do |label, parent, negate|
   with_scope(parent) do
-    expect(find_field(label)['checked']).to negate ? eq(false) : eq(true)
+    checkbox = find_field(label)
+    if negate
+      expect(checkbox).not_to be_checked
+    else
+      expect(checkbox).to be_checked
+    end
   end
 end
 
