@@ -18,15 +18,15 @@ module ActiveAdmin
       # Get's called through a before filter
       def set_current_tab
         @current_tab = if current_menu && active_admin_config.belongs_to? && parent?
-          parent_item = active_admin_config.belongs_to_config.target.menu_item
-          if current_menu.include? parent_item
-            parent_item
-          else
-            active_admin_config.menu_item
-          end
-        else
-          active_admin_config.menu_item
-        end
+                         parent_item = active_admin_config.belongs_to_config.target.menu_item
+                         if current_menu.include? parent_item
+                           parent_item
+                         else
+                           active_admin_config.menu_item
+                         end
+                       else
+                         active_admin_config.menu_item
+                       end
       end
 
     end
