@@ -5,7 +5,7 @@ module ActiveAdmin
       # Controller action that is called when submitting the batch action form
       def batch_action
         if action_present?
-          selection  =            params[:collection_selection] ||  []
+          selection  =            params[:collection_selection] || []
           inputs     = JSON.parse params[:batch_action_inputs]  || '{}'
           valid_keys = render_in_context(self, current_batch_action.inputs).try(:keys)
           inputs     = inputs.with_indifferent_access.slice *valid_keys
