@@ -10,11 +10,11 @@ module ActiveAdmin
       end
 
       private
-        def datepicker_options
-          options = self.options.fetch(:datepicker_options, {})
-          options = Hash[options.map{ |k, v| [k.to_s.camelcase(:lower), v] }]
-          { datepicker_options: options }
-        end
+      def datepicker_options
+        options = self.options.fetch(:datepicker_options, {})
+        options = Hash[options.map{ |k, v| [k.to_s.camelcase(:lower), v] }]
+        { datepicker_options: options }
+      end
     end
   end
 end
