@@ -7,9 +7,8 @@ module ActiveAdmin
         def to_html
           input_wrapping do
             [ label_html,
-              builder.text_field(gt_input_name, input_html_options(gt_input_name)),
-              template.content_tag(:span, "-", class: "seperator"),
-              builder.text_field(lt_input_name, input_html_options(lt_input_name)),
+              builder.text_field(gt_input_name, input_html_options(gt_input_name, gt_input_placeholder)),
+              builder.text_field(lt_input_name, input_html_options(lt_input_name, lt_input_placeholder)),
             ].join("\n").html_safe
           end
         end
@@ -23,7 +22,7 @@ module ActiveAdmin
           column && column.type == :date ? "#{method}_lteq" : "#{method}_lteq_datetime"
         end
 
-        def input_html_options(input_name = gt_input_name)
+        def input_html_options(input_name = gt_input_name, placeholder = gt_input_placeholder)
           current_value = begin
             #cast value to date object before rendering input
             @object.public_send(input_name).to_s.to_date
@@ -33,7 +32,16 @@ module ActiveAdmin
           { size: 12,
             class: "datepicker",
             maxlength: 10,
+            placeholder: placeholder,
             value: current_value ? current_value.strftime("%Y-%m-%d") : "" }
+        end
+
+        def gt_input_placeholder
+          I18n.t("active_admin.filters.predicates.from")
+        end
+
+        def lt_input_placeholder
+          I18n.t("active_admin.filters.predicates.to")
         end
       end
     end
