@@ -48,10 +48,10 @@ module ActiveAdmin
       #
       def add(options)
         item = if parent = options.delete(:parent)
-          (self[parent] || add(label: parent)).add options
-        else
-          _add options.merge parent: self
-        end
+                 (self[parent] || add(label: parent)).add options
+               else
+                 _add options.merge parent: self
+               end
 
         yield(item) if block_given?
 
