@@ -10,7 +10,7 @@ module ActiveAdmin
 
     before_action :authorize_access!
 
-    def index(options={}, &block)
+    def index(options = {}, &block)
       render "active_admin/page/index"
     end
 
