@@ -14,6 +14,8 @@ module ActiveAdmin
           include ScopeChain
 
           define_active_admin_callbacks :build, :create, :update, :save, :destroy
+
+          helper_method :current_scope
         end
       end
 
@@ -45,7 +47,6 @@ module ActiveAdmin
         end
       end
 
-
       # Does the actual work of retrieving the current collection from the db.
       # This is a great method to override if you would like to perform
       # some additional db # work before your controller returns and
@@ -59,7 +60,6 @@ module ActiveAdmin
         end
         collection
       end
-
 
       # Override this method in your controllers to modify the start point
       # of our searches and index.
@@ -103,7 +103,6 @@ module ActiveAdmin
         scoped_collection.send method_for_find, params[:id]
       end
 
-
       # Builds, memoize and authorize a new instance of the resource. The
       # actual work of building the new instance is delegated to the
       # #build_new_resource method.
@@ -116,7 +115,6 @@ module ActiveAdmin
         get_resource_ivar || begin
           resource = build_new_resource
           resource = apply_decorations(resource)
-          resource = assign_attributes(resource, resource_params)
           run_build_callbacks resource
           authorize_resource! resource
 
@@ -129,7 +127,7 @@ module ActiveAdmin
       #
       # @return [ActiveRecord::Base] An un-saved active record base object
       def build_new_resource
-        scoped_collection.send method_for_build
+        scoped_collection.send method_for_build, *resource_params
       end
 
       # Calls all the appropriate callbacks and then creates the new resource.
@@ -181,11 +179,9 @@ module ActiveAdmin
         end
       end
 
-
       #
       # Collection Helper Methods
       #
-
 
       # Gives the authorization library a change to pre-scope the collection.
       #
@@ -242,13 +238,15 @@ module ActiveAdmin
 
       def current_scope
         @current_scope ||= if params[:scope]
-          active_admin_config.get_scope_by_id(params[:scope])
-        else
-          active_admin_config.default_scope(self)
-        end
+                             active_admin_config.get_scope_by_id(params[:scope])
+                           else
+                             active_admin_config.default_scope(self)
+                           end
       end
 
       def apply_pagination(chain)
+        # skip pagination if already was paginated by scope
+        return chain if chain.respond_to?(:total_pages)
         page_method_name = Kaminari.config.page_method_name
         page = params[Kaminari.config.param_name]
 
@@ -297,6 +295,23 @@ module ActiveAdmin
       #
       def apply_decorations(resource)
         apply_decorator(resource)
+      end
+
+      # @return [String]
+      def smart_resource_url
+        if create_another?
+          new_resource_url(create_another: params[:create_another])
+        else
+          super
+        end
+      end
+
+      private
+
+      # @return [Boolean] true if user requested to create one more
+      #   resource after creating this one.
+      def create_another?
+        params[:create_another].present?
       end
     end
   end
