@@ -63,11 +63,16 @@ module ActiveAdmin
     def permit_params(*args, &block)
       param_key = config.param_key.to_sym
       belongs_to_param = config.belongs_to_param
+      create_another_param = :create_another if config.create_another
 
       controller do
         define_method :permitted_params do
-          params.permit *(active_admin_namespace.permitted_params + Array.wrap(belongs_to_param)),
-            param_key => block ? instance_exec(&block) : args
+          permitted_params =
+            active_admin_namespace.permitted_params +
+              Array.wrap(belongs_to_param) +
+              Array.wrap(create_another_param)
+
+          params.permit(*permitted_params, param_key => block ? instance_exec(&block) : args)
         end
       end
     end
@@ -100,7 +105,7 @@ module ActiveAdmin
     #     column :name
     #   end
     #
-    def csv(options={}, &block)
+    def csv(options = {}, &block)
       options[:resource] = config
 
       config.csv_builder = CSVBuilder.new(options, &block)
@@ -180,18 +185,24 @@ module ActiveAdmin
     delegate :before_save,    :after_save,    to: :controller
     delegate :before_destroy, :after_destroy, to: :controller
 
-    # This code defines both *_filter and *_action for Rails 4.0 to Rails 5.
-    actions = [
+    # This code defines both *_filter and *_action for Rails 4.0 to Rails 5 and  *_action for Rails >= 5.1
+    phases = [
       :before, :skip_before,
       :after,  :skip_after,
       :around, :skip
     ]
-    [:action, :filter].each do |name|
-      actions.each do |action|
-        define_method "#{action}_#{name}" do |*args, &block|
-          controller.public_send "#{action}_action", *args, &block
-        end
-      end
+    keywords = if Rails::VERSION::MAJOR == 5 && Rails::VERSION::MINOR >= 1
+                 [:action]
+               else
+                 [:action, :filter]
+               end
+
+    keywords.each do |name|
+      phases.each do |action|
+       define_method "#{action}_#{name}" do |*args, &block|
+         controller.public_send "#{action}_action", *args, &block
+       end
+     end
     end
 
     # Specify which actions to create in the controller
