@@ -46,9 +46,9 @@ module ActiveAdmin
 
     def name
       case @name
-        when String then @name
-        when Symbol then @localizer ? @localizer.t(@name, scope: 'scopes') : @name.to_s.titleize
-        else             @name
+      when String then @name
+      when Symbol then @localizer ? @localizer.t(@name, scope: 'scopes') : @name.to_s.titleize
+      else @name
       end
     end
 
