@@ -1,3 +1,3 @@
 module ActiveAdmin
-  VERSION = '1.0.0.pre5'
+  VERSION = '1.0.0'
 end
