@@ -8,8 +8,8 @@ module ActiveAdmin
         end.content
       end
 
-      def hidden_field_tags_for(params, options={})
-        fields_for_params(params, options).map do |kv|
+      def hidden_field_tags_for(params, options = {})
+        fields_for_params(params.to_unsafe_hash, options).map do |kv|
           k, v = kv.first
           hidden_field_tag k, v, id: sanitize_to_id("hidden_active_admin_#{k}")
         end.join("\n").html_safe
