@@ -5,7 +5,7 @@ module ActiveAdmin
 
       def build(action_items)
         action_items.each do |action_item|
-          span class: "action_item" do
+          span class: action_item.html_class do
             instance_exec(&action_item.block)
           end
         end
