@@ -12,7 +12,6 @@ module ActiveAdmin
       include ActiveAdmin::ScopeChain
       include ::ActiveAdmin::Helpers::Collection
 
-
       def default_class_name
         "scopes table_tools_segmented_control"
       end
@@ -31,11 +30,10 @@ module ActiveAdmin
 
       def build_scope(scope, options)
         li class: classes_for_scope(scope) do
-          scope_name = I18n.t "active_admin.scopes.#{scope.id}", default: name_for_scope(scope)
-          params     = request.query_parameters.except :page, :scope, :commit, :format
+          params = request.query_parameters.except :page, :scope, :commit, :format
 
           a href: url_for(scope: scope.id, params: params), class: 'table_tools_button' do
-            text_node scope_name
+            text_node scope_name(scope)
             span class: 'count' do
               "(#{get_scope_count(scope)})"
             end if options[:scope_count] && scope.show_count
@@ -62,12 +60,6 @@ module ActiveAdmin
         collection_size(scope_chain(scope, collection_before_scope))
       end
 
-      def name_for_scope(scope)
-        case scope.name
-          when Proc then self.instance_exec(&scope.name).to_s
-          else      scope.name.to_s
-        end
-      end
     end
   end
 end
