@@ -25,7 +25,7 @@ module ActiveAdmin
      # end
      # ```
      #
-     class IndexAsGrid < ActiveAdmin::Component
+    class IndexAsGrid < ActiveAdmin::Component
 
       def build(page_presenter, collection)
         @page_presenter = page_presenter
@@ -42,7 +42,7 @@ module ActiveAdmin
         "grid"
       end
 
-      protected
+     protected
 
       def build_table
         resource_selection_toggle_panel if active_admin_config.batch_actions.any?
