@@ -25,7 +25,12 @@ module ActiveAdmin
         end
 
         def attributes_table(*args, &block)
-          table_title = ActiveAdmin::Localizers.resource(active_admin_config).t(:details)
+          opts = args.extract_options!
+          table_title = if opts.has_key?(:title)
+                          render_or_call_method_or_proc_on(resource, opts[:title])
+                        else
+                          ActiveAdmin::Localizers.resource(active_admin_config).t(:details)
+                        end
           panel(table_title) do
             attributes_table_for resource, *args, &block
           end
