@@ -29,10 +29,10 @@ module ActiveAdmin
         breadcrumb_config = active_admin_config && active_admin_config.breadcrumb
 
         links = if breadcrumb_config.is_a?(Proc)
-          instance_exec(controller, &active_admin_config.breadcrumb)
-        elsif breadcrumb_config.present?
-          breadcrumb_links
-        end
+                  instance_exec(controller, &active_admin_config.breadcrumb)
+                elsif breadcrumb_config.present?
+                  breadcrumb_links
+                end
         return unless links.present? && links.is_a?(::Array)
         span class: "breadcrumb" do
           links.each do |link|
