@@ -46,19 +46,26 @@ module ActiveAdminIntegrationSpecHelper
   # Returns a fake action view instance to use with our renderers
   def mock_action_view(assigns = {})
     controller = ActionView::TestCase::TestController.new
+    #this line needed because of rails bug https://github.com/rails/rails/commit/d8e98897b5703ac49bf0764da71a06d64ecda9b0
+    controller.params = ActionController::Parameters.new
     MockActionView.new(ActionController::Base.view_paths, assigns, controller)
   end
   alias_method :action_view, :mock_action_view
 
   # A mock action view to test view helpers
   class MockActionView < ::ActionView::Base
-    include ActionView::Helpers
     include ActiveAdmin::ViewHelpers
     include Rails.application.routes.url_helpers
   end
 
   def with_translation(translation)
-    I18n.backend.store_translations :en, translation
+    # If no translations have been loaded, any later calls to `I18n.t` will
+    # cause the full translation hash to be loaded, possibly overwritting what
+    # we've loaded via `store_translations`. We use this hack to prevent that.
+    # @todo Might not be necessary anymore once
+    # https://github.com/svenfuchs/i18n/pull/353 lands.
+    I18n.backend.send(:init_translations)
+    I18n.backend.store_translations I18n.locale, translation
     yield
   ensure
     I18n.backend.reload!
