@@ -1,8 +1,10 @@
 # Rails template to build the sample app for specs
 
-if Rails::VERSION::MAJOR == 4 && Rails::VERSION::MINOR >= 2
-  copy_file File.expand_path('../templates/manifest.js', __FILE__), 'app/assets/config/manifest.js', force: true
-end
+copy_file File.expand_path('../templates/manifest.js', __FILE__), 'app/assets/config/manifest.js', force: true
+
+create_file 'app/assets/stylesheets/some-random-css.css'
+create_file 'app/assets/javascripts/some-random-js.js'
+create_file 'app/assets/images/a/favicon.ico'
 
 generate :model, 'post title:string body:text published_date:date author_id:integer ' +
   'position:integer custom_category_id:integer starred:boolean foo_id:integer'
@@ -12,7 +14,7 @@ create_file 'app/models/post.rb', <<-RUBY.strip_heredoc, force: true
     belongs_to :author, class_name: 'User'
     has_many :taggings
     accepts_nested_attributes_for :author
-    accepts_nested_attributes_for :taggings
+    accepts_nested_attributes_for :taggings, allow_destroy: true
 
     ransacker :custom_title_searcher do |parent|
       parent.table[:title]
@@ -41,7 +43,7 @@ create_file 'app/models/blog/post.rb', <<-RUBY.strip_heredoc, force: true
     belongs_to :author, class_name: 'User'
     has_many :taggings
     accepts_nested_attributes_for :author
-    accepts_nested_attributes_for :taggings
+    accepts_nested_attributes_for :taggings, allow_destroy: true
 
     if defined? ProtectedAttributes
       attr_accessible :title, :body, :starred, :author, :position, :published_date, :author_id, :custom_category_id, :category
@@ -99,33 +101,22 @@ RUBY
 
 generate :model, 'store name:string'
 
-# Generate a model with string ids
 generate :model, 'tag name:string'
-gsub_file Dir['db/migrate/*_create_tags.rb'].first, /\:tags do .*/, <<-RUBY.strip_heredoc
-  :tags, id: false, primary_key: :id do |t|
-    t.string :id
-RUBY
 create_file 'app/models/tag.rb', <<-RUBY.strip_heredoc, force: true
   class Tag < ActiveRecord::Base
-    self.primary_key = :id
-    before_create :set_id
-
-    private
-    def set_id
-      self.id = SecureRandom.uuid
-    end
-
     if defined? ProtectedAttributes
       attr_accessible :name
     end
   end
 RUBY
 
-generate :model, 'tagging post_id:integer tag_id:integer'
+generate :model, 'tagging post_id:integer tag_id:integer position:integer'
 create_file 'app/models/tagging.rb', <<-RUBY.strip_heredoc, force: true
   class Tagging < ActiveRecord::Base
     belongs_to :post
     belongs_to :tag
+
+    delegate :name, to: :tag, prefix: true
   end
 RUBY
 
@@ -133,10 +124,12 @@ gsub_file 'config/environments/test.rb', /  config.cache_classes = true/, <<-RUB
 
   config.cache_classes = !ENV['CLASS_RELOADING']
   config.action_mailer.default_url_options = {host: 'example.com'}
-  config.assets.digest = false
+  config.assets.precompile += %w( some-random-css.css some-random-js.js a/favicon.ico )
 
-  if Rails::VERSION::MAJOR >= 4 && Rails::VERSION::MINOR >= 1
-    config.active_record.maintain_test_schema = false
+  config.active_record.maintain_test_schema = false
+
+  if Rails::VERSION::MAJOR >= 5
+    config.active_record.belongs_to_required_by_default = false
   end
 
 RUBY
@@ -169,10 +162,6 @@ directory File.expand_path('../templates/policies', __FILE__), 'app/policies'
 if ENV['RAILS_ENV'] != 'test'
   inject_into_file 'config/routes.rb', "\n  root to: redirect('admin')", after: /.*routes.draw do/
 end
-
-# Devise master doesn't set up its secret key on Rails 4.1
-# https://github.com/plataformatec/devise/issues/2554
-gsub_file 'config/initializers/devise.rb', /# config.secret_key =/, 'config.secret_key ='
 
 rake "db:drop db:create db:migrate", env: 'development'
 rake "db:drop db:create db:migrate", env: 'test'
