@@ -1,4 +1,3 @@
-require 'spec_helper'
 require 'active_admin/menu'
 require 'active_admin/menu_item'
 
@@ -97,7 +96,7 @@ module ActiveAdmin
 
       context "with no parent" do
         it "should return an empty array" do
-         expect(item.ancestors).to eq []
+          expect(item.ancestors).to eq []
         end
       end
 
@@ -127,7 +126,6 @@ module ActiveAdmin
         end
       end
     end # accessing ancestory
-
 
     describe "#id" do
       it "should be normalized" do
