@@ -2,7 +2,7 @@ require 'rails_helper'
 
 def it_should_have_view(key, value)
   it "should have #{value} for view key '#{key}'" do
-    expect(subject.send(key)).to  eq value
+    expect(subject.send(key)).to eq value
   end
 end
 
