@@ -4,7 +4,9 @@ RSpec.describe ActiveAdmin::Views::Pages::Form do
   describe "#title" do
     let!(:application){ ActiveAdmin::Application.new }
     let(:namespace){ ActiveAdmin::Namespace.new(application, "Admin") }
-    let!(:params){ { controller: "UsersController", action: "edit" } }
+    let!(:http_params){ { controller: "UsersController", action: "edit" } }
+    let!(:params) { ActionController::Parameters.new(http_params) }
+
     let(:helpers) do
       helpers = mock_action_view
       allow(helpers).to receive(:active_admin_config).and_return(namespace.register(Post))
@@ -18,6 +20,7 @@ RSpec.describe ActiveAdmin::Views::Pages::Form do
 
     context "when :title is set" do
       it "should show the set page title" do
+
         page = ActiveAdmin::Views::Pages::Form.new(arbre_context)
         expect(page).to receive(:resource)
         expect(page).to receive(:form_presenter).twice.and_return({ title: "My Page Title" })
